[FEAT] : ⚙️: 인터페이스 수정
