+++ app/src/main/java/com/example/ymd/retrofit/Constans.kt
@@ -1,0 +1,17 @@
+package com.example.ymd.retrofit
+
+object Constans {
+
+    // youtube  API의 기본 URL입니다.
+    const val BASE_URL = "https://www.googleapis.com/youtube/v3/"
+
+    // youtube API를 사용하기 위한 인증 헤더입니다.
+    const val AUTH_HEADER = "AIzaSyBaTftuay-7bov4muIG4oeVRtrHJ4E15FU"
+
+    // 앱의 Shared Preferences 파일 이름입니다.
+    const val PREFS_NAME = "com.example.ymd.prefs"
+
+    // 마지막 검색어를 저장하기 위한 키 값입니다.
+    const val PREF_KEY = "VIDEO_SEARCH_PREF"
+
+}
